remove old artworks component and fix up relay fragment
